requirements.txt
ultralytics
roboflow
opencv-python
numpy
easyocr
Pillow
streamlit

